+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # QOwnNotes Changelog
 
-## 23.5.4
+## 23.6.0
 - added more Dutch translation (thank you stephanp)
 - the QOwnNotes command-line snippet manager [qc](https://github.com/qownnotes/qc)
   can now be installed via the [Nix Store](https://search.nixos.org/packages?channel=unstable&show=qc)
   - you can use `nix-shell -p qc --run qc` to try it out
+- the Korean translation was now integrated into the webpage on
+  [QOwnNotes webpage in Korean](https://www.qownnotes.org/ko)
+  (for [#2788](https://github.com/pbek/QOwnNotes/issues/2788))
 
 ## 23.5.3
 - added more Arabic translation (thank you noureddin)